Support lights with sprite particles
Add support for lights with the sprite particles. Add API to define
all lights which particle should use. By default this is empty, so
default behavior remains the same (lights not affecting the particles).
Lights are only calculated on vertex shaders, for performance reasons
and because particles are usually small in size. Lights have own shader
feature levels to not affect the performance of non-lighted particles.

Includes an example and lancelot tests.

Task-number: QTBUG-90797
Change-Id: Idca84f6ec73e36d483c0598a9c08a54cec238afa
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>