Add support for KHR_materials_ior to PrincipledMaterial
[ChangeLog][Quick3D][QQuick3DPrincipledMaterial] This patch changes the
default value of specularAmount to 1.0. This property as documented
should only modify the amount of the specular contribution. A value of
1.0 means that full specular contribution, where a value less than 1.0
mean less than the full specular contribution. Previously the value of
0.5 was halfing the specular contribution leading to much less
impressive reflections. This was because the value of specularAmount was
also being used under the hood to calculate the dielectric specular
value used for the F term. It is much more accurate to use the index of
refraction instead for this.

Change-Id: Ia73844b5cd12a20b951eaf7825cf5de0ddcb1275
Reviewed-by: Christian Strømme <christian.stromme@qt.io>