Update assimp to version 5.1.1
Assimp finally had a decent release which we can support as a system
version where possible.

For the system assimp dependency we have to check for >= 5.1 and not >=
5.1.1 because the cmake files released did not define the patch release
number correctly.  So while we don't support 5.1.0 (5.1.1 contains a
patch specifically for Qt), the configure test will fail if you try and
build with 5.1.0, so this should be fine.

Change-Id: Ie7bf849d9e53ea1eb10f9099f235b18c2f52caf0
Reviewed-by: Christian Strømme <christian.stromme@qt.io>