Apply Occlusion map to global ambient light as well
Previously use of occlusion maps (baked AO) was limited to materials
that took advantage of IBL. That is because AO applies to indirect light
and not direct light. Punctual lights do provide a way to simulate
indirect light via the ambientColor property, which means we should also
apply AO from occlusion maps to this light as well.

[ChangeLog][Quick3D][QQuick3DPrincipledMaterial] Occlusion maps now also
affect the indirect lighting provided by punctual light's ambientColor

Pick-to: 6.3 6.2
Change-Id: I71016da696ecc1d7ad5e0e5de82b13ad49077d52
Reviewed-by: Inho Lee <inho.lee@qt.io>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>