QQuick3DViewport: remove a use of QMutableEventPoint
'eventPoint' is never used as a QMutableEventPoint, so don't cast to
it (which is UB).

Also make it a reference. The effect is the same, thanks to explicit
sharing, but the rest of the code in Qt uses references, so do it
here, too.

Task-number: QTBUG-99615
Change-Id: Ib2b756c30a9fc775f1f3541583ff6385d6f59282
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 95aac86e683dafd0a43a1e7172edc21b0648f592)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>