Skinning : add Skin type for better animation
Skeleton type is an ambiguous type. This patch introduces a new
resource for skeletal animations, Skin. It's a pure resource type and
any Node type can be used for the skeleton heirarchy.

Model can have a Skin for the animation and it has a higher priority
than Skeleton. Skin has a list of joints and an optional list of
inverse bind matrices.

Skin {
    id: skin0
    joints: [
        joint0,
        joint1
    ]
    inverseBindPoses: [
        Qt.matrix4x4(1, 0, 0, -0.5,
                     0, 1, 0, -1,
                     0, 0, 1, 0,
                     0, 0, 0, 1),
        Qt.matrix4x4(1, 0, 0, -0.5,
                     0, 1, 0, -1,
                     0, 0, 1, 0,
                     0, 0, 0, 1)
    ]
}

Task-number: QTBUG-93361
Change-Id: I52b074abda8d75451308f3f23abfac8a8bd1411e
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>