Make sure to convert non HDR light probe images to linear color space
Currently if you try and use a non HDR image as a light probe, we don't
convert it to linear color space. This means that that the skybox will
get try and convert an sRGB image to sRGB color space leading to a very
washed out appearance. Now we will make sure to do the color space
conversion when we create the environment map if needed.

Fixes: QTBUG-94616
Pick-to: 6.3 6.2
Change-Id: Icc1111a3815252b009ebc95f0bdd925ed6e0d05a
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>