PrincipledMaterial Example Overhaul
This is a complete re-write of the PrincipledMaterial example. The
previous example really only scraped the surface of what
PrincipledMaterial is, and part of what it did show off was more of a
specialized usage.

This re-write is an attempt to not only show off the
PrincipledMaterial's features, but also educate the end users how each
of the properties work as well as enable experimentation with each of
the features.

Change-Id: I940f8c9419fc15c5d30509e863d2838e06316c64
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit ef72d4beccf0f1ea03ea95176be070f1b7e13c02)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>