Always fallback to Red if requested Texture Channel is not present
Previously the code tried to fallback to predetermined channels, but
only 5 defaults were defined in this list, and these defaults ignored
the which channels actually existed in the texture. Having a fallback
list also ignored the fact that the defaults of each channel are already
defined and documented by the material.  If a user overrides this
default, and the channel doesn't exist that is the users fault. We
should not try to fallback to the original default, instead we should
fallback to the only channel that will always exist if the texture
exists, which is the first channel: Red.

Pick-to: 6.3
Change-Id: Ia827f391fab45ca77c504aadb1d3a46478a7b38c
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
Reviewed-by: Inho Lee <inho.lee@qt.io>