Use QMatrix4x4's normalMatrix for calculating the normal matrix
It does have some benefits over the getInverse(), which can now be
removed because it is not used for anything but to get a normal
matrix in various places. The QMatrix4x4 implementation uses double
precision internally for the determinant, which turns out to be
important with certain transformation matrices from scenes with models
with small scale factors.

Pick-to: 6.2 6.3
Change-Id: Id1443adb9f0f21af425bada015af7e88bbf41445
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>