Disable all tonemapping in Screen Texture pass
The rendering assumes that the contents of the screen texture is in
linear color space, but since it uses the same materials from the
regular color pass if tonemapping is enabled then it will be enabled in
the screen texture pass.  The Skybox (Clear) pass is already bypassing
tonemapping, so the situation now is that there is already a mixed
content texture. This change makes it where the objects drawn in the
screen texture are also not tonemapped.

I expect this to break some lancelot tests involving Transmission.

Change-Id: Iaad892985d9a8b56e3f1808aa5fd2f188b258381
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 99a3b73b4e30f49a910af5be24df5f478a4f3c17)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>