Doc: ResourceLoader usage in PrincipledMaterial Example
ResourceLoader is pretty hard to example on its own because it about
being explicit about GPU resource lifetimes. So in this case I document
the usage of ResourceLoader inside of the PrincipledMaterial example
since it is used with purpose to pre-load and preserve resources. I
think this is better than creating a contrived example specifically for
the resource manager.

Task-number: QTBUG-99556
Pick-to: 6.3
Change-Id: Ie7aa08ae844c3ebea05583a263a142d9006030e2
Reviewed-by: Christian Strømme <christian.stromme@qt.io>