Correct probeExposure behavior
2 issues should be resolved:
1) Exposure mapping should only happen when tonemapping is enabled, as
it is part of the tonemapping. The reasons it was happening inline in
materials is to avoid a post processing step for tonemapping, but if
tonemapping is disabled we should assume that the user will do their own
tonemapping in a post process effect and we shouldn't touch the raw
values provided by the light probe.
2) While the Skybox shader was doing inline exposure mapping correctly,
materials were just treating exposure as if it was brightness. This lead
to the material lighting not matching the skybox. Now instead we do
exposure mapping inline when necessary when sampling values from the
light probe.

Task-number: QTBUG-96624
Change-Id: I27155b9ebfa4d02ce6b80cdfa733ccdaa30b6358
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 64c9f1d6158aaa487d0fe2fbeaed5dc40fb4410d)
Reviewed-by: Andy Nichols <andy.nichols@qt.io>