Add cubemap support for custom material/effect textures
One can now use CubeMapTexture, a Texture subclass, with a .ktx
file with 6 faces. Using the same .ktx file is still possible with
Texture as well, one just gets face 0 (X+) then, and a sampler2D in
the shader, like before. When switching to CubeMapTexture, the same
source asset now loads all 6 faces and gives a samplerCube in the
shader.

Using a CubeMapTexture is only possible with CustomMaterial and Effect
at the moment. Attempting to use it with a Default/PrincipledMaterial
will fail since those continue to treat everything as 2D.

Sourcing data via other means is not currently supported for
CubeMapTexture, meaning one can only use the source property with it.

Specifying a non-cubemap image as CubeMapTexture's source is functional,
as long as the width and height of the image are equal, but this is
intentionally not documented and advertised, because what it leads to is
creating a  cubemap texture where only face 0 is filled with data, the
rest of the faces have undefined content (and we cannot easily implement
even clearing the remaining faces to something, think compressed formats)

This also provides the enablers for easily adding 3D (volume) texture
support later on.

[ChangeLog][QtQuick3D] Added CubeMapTexture, a Texture subtype that
allows exposing a cubemap from a KTX container to custom material and
effect shaders.

Task-number: QTBUG-101554
Change-Id: Ie7e11d10c2c1602c8f564027981f82285e953051
Reviewed-by: Andy Nichols <andy.nichols@qt.io>