Add per-model lightmap UV unwrapping
Unwrapping happens now once, on the complete, merged vertex data for the
whole model, including all subsets.

This results in a single lightmap size hint. That is still stored
per-subset, we will not be changing the mesh file format just for this,
so the value is simply set to the same value for all subsets.

Change-Id: Ic17e15397370fd176c79c30b36563c6689fe9c57
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>