Renderer: Remove m_lastFrameLayers and just clear before use
This is probably all legacy stuff that doesn't really make as much sense
since we don't recursively render layers anymore. In practice this is a
reference to the QSSGLayerRenderData and there should likely only ever
be one in the list. Instead of adding weird logic to register layerdata
with the renderer, just always clear the layerdata before we use it.
Only the first time, this may be a bit redundant, but as everything it
is clearing is already empty it should be quick anyway.

Change-Id: I3f2128bf821df31d23ef5f28bf9692e373116633
Reviewed-by: Christian Strømme <christian.stromme@qt.io>