Cleanup QSSGLayerRenderData and QSSGLayerRenderPreparationData
The filenames containing the definitions of these classes have been
changed to reflect the actual name of the class (as is our custom).
The "_rhi" posfix has been removed from the filename for
QSSGLayerRenderData since it really doesn't mean anything anymore (RHI
usage is everywhere now).
Both QSSGLayerRenderData and QSSGLayerRenderPreparationData are now
classes instead of structs.

Change-Id: Ie7bdc23cbdbbe9fa195a6ffc65c24b64a500ce44
Reviewed-by: Christian Strømme <christian.stromme@qt.io>