Use map() not mapVector() in recently changed code
Some of these must be incorrect because Lancelot shows regressions
with particles. In general, the correct approach is to call map(),
that is also what the now-deprecated operator* did. In some cases,
mapVector() can be correct, depending on the context around the value
in question (is it treated as a point? or a direction vector?), but
when migrating away from the deprecated operator* just using map()
everywhere is a safe choice.

Change-Id: If00b1272caba3d4fde1bd651b9dfff597e0d8a7e
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>