Improve GLSL shader version logic
GLSL 140 is the ideal minimum, not 130, because features like instancing
use GLSL functions such as inverse that are not in 130.

For some drivers this wouldn't matter because they do not care that much
about the #version declared in the shader code. For others it might.

None of this applies if a core profile context was chosen, which is the
case when the application is well-behaved and takes idealSurfaceFormat()
into account.

The logic for choosing what GLSL version to request from SPIRV-Cross is
complicated. Originally the baker setup function did not look at the
OpenGL context the QRhi was using and rather decided based on the
defaultSurfaceFormat(). This was fine as long as an application was
calling idealSurfaceFormat() and in main() and setting the returned
format as the default.

But when that is not done, we got 130 shaders (with non-GLES) even if
the context we got was something much higher (e.g. 4.6 compatbility
profile), which is not sufficient for strict GLSL compilers and features
that involve generating shader code with 140+ GLSL functions.

We still need to have a way to run, as much as possible, with OpenGL 3.0
and GLSL version 130 however, because that's what old Mesa llvmpipe
builds provide, and that includes our "official" opengl32sw.dll shipped
with Qt. So generate GLSL 130 in that case.

In short, we now look at the OpenGL context the QRhi claims to be
using, and base the decision on that, not
QSurfaceFormat::defaultFormat(). The benefit here is that applications
that do not use idealSurfaceFormat() in their main() get better defaults
and won't break with certain drivers when using certain features (when
running with non-ES OpenGL).

For the offline shader generation this is not an option, there 130 is
bumped to 140.

Fixes: QTBUG-101638
Change-Id: I768c868c04fbf11defead02addb70fc73e2bef9c
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>