Update QQ3D morph animation to use a vertex texture
Previous implementation has a serious restriction related to the number
of input attributes. Now we moved it to use the vertex texture.
Note that it is not supported by OpenGL ES 2.0.

For a custom geometry, all the target attributes should be added by
addTargetAttribute and the data should be added by setTargetData.

This patch includes internal implementation to expand supported
target attributes to texture and color but another patch will
be followed to support them.

Fixes: QTBUG-101671
Change-Id: I8781a80d40481c4776d0511297e6b099dd29592c
Reviewed-by: Andy Nichols <andy.nichols@qt.io>