Normalize HeightFieldGeometry and fix orientation
Map the image x-axis to the scene's x-axis, and let the height
field be non-mirrored when seen from above.

Let the dimensions of HeightFieldGeometry be independent of the
resolution of the image, and let the origin be the mid point.
This means that a HeightFieldGeometry will get the same size and
position as a #Cube by default.

Rename heightFieldScale to extents, and let the default extents
preserve the aspect ratio of the image.

Also add an option to turn off smooth shading.

Change-Id: Ia7674206b6e7f9f9bf70c7680f32cf8fbcbe5914
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>