Deduplicate camera distance based sorting key generation
It is repeated in three places, and there may be more in the future -
move the calculation of the squared distance to a helper function.

Also remove an unused argument for getOpaque... - the sorting is always
desired so don't bother with the option.

Add the 'sorted' word to all the "getters".

Change-Id: I8ecf85a3e4806a95c55440478878eb9d26ca28b4
Reviewed-by: Andy Nichols <andy.nichols@qt.io>