Add Embree
Builds only the minimal set of functionality we need, which is
triangle mesh geometry building and ray intersection tests.

This is a dependency for the lightmapper, which is effectively a
desktop feature. Lightmaps are to be baked on powerful desktop PCs
running Windows, macOS, or maybe Linux, and that is when raytracing,
and so Embree, is involved. Using the generated data at run time is a
different story, that won't need Embree. Thus the build of 3rd party
bundled library is limited to:
- x64 Windows
- x64 or arm64 macOS
- x64 or arm64 Linux

The Embree configuration is limited to SSE2 (or SSE2 emulation via
NEON for ARM, AFAICS) to prevent problems (like getting binaries not
universally working everywhere). This can be revisited later, but the
expectation is that this is good enough for our purposes when it comes
to performance.

Additional notes:

- ARM64 Windows is not currently supported as Embree fails to build.

- 32-bit ARM should not be very relevant for lightmap baking purposes,
  which is good because it needs to be disabled as well because Embree
  fails to build in some of the CIs armv7 configs.

- WebAssembly has not been investigated.

Change-Id: I70211a03f66b941922d4ad04e3b64559e7cb2253
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>