Fix Point/Spot light shadows to work in all directions
For some reason the shader was inverting the z position of the camera
before calculating the depth. This makes no sense, because the idea is
to get the distance between the light and the fragment we are shading,
so by inverting a single field of a translation, we completely change
where we could never be testing the correct value along the z axis.

Fixes: QTBUG-102023
Change-Id: I3051435c64c3810348a3787c7193a1f0df3f4d03
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit df5bd7cf9ab61217107e7bd6e0bc3da30f05c595)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>