Update commercial license headers
Updated header.COMM to the files in tqtc-qtquick3d.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4968
Change-Id: Iba27bf3ac0721f54e9edde5d743b8ab2b1cbf959
Reviewed-by: Andy Nichols <andy.nichols@qt.io>