Extract the dirty skeleton updating out of the node collection code
The search and collecting of nodes should be as simple as possible for
better optimization, readability and ease of modifying. The updating of
the skeleton nodes can easily be done in a separate step which also
make it easier to reason about what the code should be doing.
Note that the dirty skeleton nodes are now kept in a map/set
instead of a list as the primary usage here is look-ups with an unknown
amount of nodes.

Change-Id: Ic257c275654aed617231b772d6de7028f35d61b7
Reviewed-by: Inho Lee <inho.lee@qt.io>