Calculate the distance from the camera when adding renderables
In retrospect it was a mistake to make the distance property optional
when creating the sorting data, as we in practice left that as a
separate step instead of calculating directly when inserting, which
meant iterating over the item an extra time and modifying its content
in the process. This is now avoided. In addition we pass the camera
data (distance and position) to all the prep function to avoid querying
for every renderable.

Change-Id: I72417c886a5914a912b5f6cf29d55b537dcd342c
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>