Add SpecularGlossyMaterial
This is basically PrincipledMaterial using the Specular/Glossiness
workflow instead of Metalness/Roughness. The DefaultMaterial generator
is still used on the backend but has some new plumbing to use the
specularTint, specularMap properties as the specular contribution and
resuses the roughness properties as glossyness. Albedo is used instead
of BaseColor to indicate that the color value does not include
reflectance.

SpecularGlossyMaterial does not have an index of refraction property
like PrincipledMaterial because this is baked into the specular color in
the Specular/Glossy workflow.

The PrincipledMaterial example has been extended to be able to switch to
Specular/Glossy mode since there is a large overlap between between
theses two materials, and technically they are both "principled" by
design.

Change-Id: I2117550669f103635aa5ec09f6b3542941101de6
Reviewed-by: Christian Strømme <christian.stromme@qt.io>