doc: Improve lightmap page formatting
The tables with two columns do not work well with larger images.
While we are reducing the image sizes, it is probably the best
to get rid of the tables altogether.

Also mention the qlmdenoiser repo, which is public, as it is unlikely
that we will get any other solution for 6.4.

Task-number: QTBUG-103340
Change-Id: I1bf8e4cb6de8b63ef0ab73b9644ee6f17ace5acb
Reviewed-by: Andy Nichols <andy.nichols@qt.io>