Add persistent storage for lightmap UV enabled meshes
So now the lightmapper also writes out a qlm_[key].mesh file in addition
to qlm_[key].exr. Unlike the exr, this one is optional. If the
application does not ship it or deletes it right away or if for any
reason it is not avilable at run-time, that's fine, the lightmap UV
generation / mesh update will happen at run-time. (for custom geometry
or built-in primitives this is the only choice anyway)

But, with big models, take Sponza for example, it is not great if the
application will need to reprocess the whole mesh every time at scene
load just because we have a lightmap associated with the model. So in
this case the application can choose to also ship the qlm_*.mesh file
next to the .exr one, in which case it will get picked up automatically.

Task-number: QTBUG-103340
Change-Id: Ifcb7bd364b800c62fe025aa35e3d5211bb95abbb
Reviewed-by: Andy Nichols <andy.nichols@qt.io>