Fix how texture input is tracked in the custom material and the effect
The effect and custom material can only reasonably track the texture
inputs (changed/destroyed), the texture inputs are also unique and cannot
be used with more then one uniform, which make it easier to track them
in the effect and material since more then one input can map to the same
texture. Previously, if a texture was shared between multiple uniforms/
texture inputs and a texture input was destroyed we'd keep a handle to
the texture forever, as we had no way to know when it no longer was in
use.

Change-Id: I03896b3aaebbe94fb2aa0d4426e526787ba757d6
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>