Improve sorting/priority of dirty nodes
- Added more structured and extensible logic for handling updating of
dirty items.
- preSync() is now called for all scene objects, not just nodes and the
ones that happened to end up in the "right" dirty node list.
- Lights are not kept in a separate dirty list or processed twice, but
instead processed after all other nodes. A check is also added to check
if a light is scoped to itself.
- Skeleton nodes are processed as nodes, just with a higher priority
which is more sensible then just throwing it in the same pile as the
resource objects.

Change-Id: I05c13b07adc245fc1de48e97c3f57e2fa6418b39
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>