materialeditor: Change preview from ColumnLayout to items on top
What we need is the preview controls on top of the View3D. It is not
clear why ColumnLayout worked as it did in 6.3, because in
6.4 and dev the preview controls do no show up, they are presumably
outside the view at the bottom, as is expected with a ColumnLayout
if the View3D fills all the space. It could also be some change
around the Layout.preferredWidth/Height properties. Avoid all this
by not using a layout.

Pick-to: 6.4
Change-Id: I8c1c5fa043bc6a7947fcbbe568db4412e231fe98
Reviewed-by: Christian Strømme <christian.stromme@qt.io>