Split effect shader prep into two
This does not change the visible behavior yet in any way. It allows
taking layer (SceneEnvironment) settings into account, the prime
example of which is going to be the tonemapping mode. To enable this,
some of the work that was previously performed in
QQuick3DEffect::updateSpatialNode() is moved to the backend object and
triggered only in QQuick3DSceneRenderer::updateLayerNode() after the
SceneEnvironment settings are synced and the postprocessing effect list
the renderer is going to use is finalized. The latter allows
implementing logic where something is done only in the first or the last
effect in the chain.

Task-number: QTBUG-104759
Change-Id: I63eb5ac62907eb5e0c07f417709aed2f683d47b8
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>