Fix type checks introduced in the priority change
This amends 55e10112e1855 which contains the wrong type checks when
going over the nodes and resources. The logic should be that all nodes
are identified as a node type (verified when object is created), all
other objects are assumed to be some for of resource, even if they
don't have a proper type set (Unknown type), which is why we cannot
have the assert based on the type being a resource or not.

Change-Id: I84ca9ecc5233363b2f6103faa60295818bbf5078
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>