Enable proper tonemapping for effects
On its own this patch is sufficient to get proper results with a
single effect in the environment's effect list, or with multiple ones
in case the effect list is static enough. However, dynamically altered
effect lists need more dynamic behavior, as all effects need to be
able to reevaluate certain things then (like what is now the last
effect when it comes to tonemapping, and perhaps update the shaders),
but effects miss this completely: they do not even react when the
shader source properties change and lack any sort of dirty mechanism
for non-user properties. Therefore, this is going to be fixed in
separate patches only.

The last effect's last pass will now perform automatically the same
tonemapping as the main render pass would, depending on the
SceneEnvironment's tonemapMode. If it is TonemapModeNone, there is no
visual change in the result.

There is another problem to be fixed separately: consider a scene with
a given background clear color. The main pass's tonemapping does not
tonemap the background (as the color just gets passed to the QRhi as
the render pass's clear color; there is no srgb->linear conversion
either which all works if the tonemap mode is the default Linear but
incorrect if the mode is anything else). When run through an effect,
that can only operate on a texture with all the content, background
included, so now it will tonemap the background pixels too.

A skybox should work as expected however, because there we can apply
the same logic as in the main pass, i.e. use the non-tonemapped skybox
shaders when there is an effect present, so no problem there.

So the case of the Color background mode has to be fixed separately,
otherwise the output from an effect will lead to a background color
different from what's shown when there are no effects in the scene.

Task-number: QTBUG-104759
Change-Id: Icd2ea78c659cd0ba3dbbc0a2fe10db3db64b1841
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>