Tonemap the clear color
The color needs to go through sRGB->linear like all other colors.

Then based on the tonemapMode it needs to be converted as we would do
for a fragment color in the material or skybox shaders.

With the default tonemap mode Linear this brings no visible change,
whereas the other tonemap modes will now have a proper effect on the
clear color (even None, because now we treat the color as sRGB like
with all other color properties, meaning no tonemapping implies
passing in a linearized value to the 3D API).

This also means that post-processing effects, when they perform the
built-in tonemapping in the last effect, will now output an image
where the background looks as expected, i.e. not different from when
there is no effect, because all code paths perform tonemapping
correctly now.

Fixes: QTBUG-104885
Task-number: QTBUG-104759
Change-Id: Ia2bba81dbaecb09c1836314e9861ba619ae36b37
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>