Make effects update themselves when the chain changes
Appending to the SceneEnvironment's effects list triggers an update
on all the effects. Which is particularly important when the last
effect is now something else, which means the previously last one does
not need tonemapping anymore whereas the new last one does.

CustomMaterial and Effect are now more similar when it comes to the
updateSpatialNode(). This also removes an odd repeated signal connection
in custom material which should not be necessary.

This also enables handling other types of changes in the future, such as
changes in the shader source uris.

Fixes: QTBUG-104759
Change-Id: I148d6ee34ff54cdeaf3e15ad034534863b6f44c3
Reviewed-by: Andy Nichols <andy.nichols@qt.io>