Adjust InstanceRepeater documentation
Don't claim that it is impossible to implement instanced picking
directly when we already have an implementation for Qt 6.5. Include
an optimization hint which will still be relevant when the new picking
implementation is in place.

Change-Id: I9815fc0685956de4e39dbf101677fde1933f8908
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>
(cherry picked from commit aebf2038fc3685d8948a0e6636589a83f7c2a14c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>