Set the render nodes to be active by default
The initial idea was to avoid doing global updates for nodes that were
not active. But if we're going to move when the global updates are
done anyways it probably more reasonable to assume nodes are active
by default (locally). Having the node active by default also means
less frustration when nodes created "manually" seemingly don't show up.

Change-Id: Ie1f04913e06076ac984ace9cf7363e95cb546cb4
Reviewed-by: Andy Nichols <andy.nichols@qt.io>