Use FrameAnimation for the OrbitCamera and Wasd Controllers
Both of these classes were still using a 16ms timer to poll updates,
which leads to a not so smooth experience. Instead now we use the
FrameAnimation class to always update when frames are being rendered.
This also now takes into consideration the actual frame delta's to
further smooth the animations giving actual meaning to the speed values.
More realisticly the speed values are meters per second, assuming the
default unit scale is cm, otherwise it's just units per second times *
100.

OrbitCameraController and WasdController should be smooth as 🧈 with this
patch.

Pick-to: 6.4
Change-Id: I3e00d2555b38b06829f31a8cf544e16dd0d6bd10
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>