dynamic texture example: use exclusive grab on fallback TapHandler
When the user successfully clicks a TextEdit in a sticky note on a door to
set the cursor position, the TextEdit takes an exclusive grab. But there's
also a TapHandler on the View3D that opens or closes the door when tapped
(that's not the only way to do it, but it provides an opportunity to
demonstrate picking in QML). It gets the grab first because it's
"on top". So it works better to have TextEdit take over the exclusive
grab so that the TapHandler gets canceled, rather than relying on
TapHandler to explicitly check whether some other item has a grab and
whether that other item is in a subscene, got its grab directly vs. by
filtering, or whatever. When TextEdit gets the event after TapHandler
gets it, it tries to take the grab unconditionally, and TapHandler's
default grabPermissions allow stealing (ApprovesTakeOverByAnything).
(If we wanted the TapHandler to "win", there are at least two ways: let
it use a passive grab, or change grabPermissions to prevent stealing.)

The previous fix for QTBUG-94012 seems inconsistent with the idea that
passive-grabbing handlers should be able to override behavior of other
components; making up niche rules tends to break other use cases.

Task-number: QTBUG-94012
Task-number: QTBUG-105609
Change-Id: I80229510f92312f1780d96e562472e2cfee98e71
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 29a2f67a85ed4c5c14dd59972468d481e1e87104)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>