Quick3DUtils: Fix bug in RenderData's compare function
We were not considering that the divergence from 1.0 (1.0 - dot)
could produce negative values, which again would cause the wrong values
to be accepted when testing against the fudge factor in the
internal fuzzy quat compare.
And the error value was changed into 1e-7 in order to get more precision
than qFuzzyCompare of QQuaternion.

Fixes: QTBUG-105918
Pick-to: 6.4
Change-Id: I266d388e6a1bf98bb257e5d9165573810675541b
Reviewed-by: Christian Strømme <christian.stromme@qt.io>