Fix Reflection Probe Debug View
If View3D is the root of the QML file, the
Reflection Probe will fail to find it because
it starts searching from the children of the
root item. This fix makes the probe to start
searching from the root item.

Fixes: QTBUG-106037
Change-Id: I3e33d30134e189394759f804731290483f49d3f1
Reviewed-by: Antti Määttä <antti.maatta@qt.io>
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
(cherry picked from commit 8042f6d5d997526b76a1d8b626be01459b396fa9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>