Encapsulate passes
This is the first step in getting a more flexible and easy to
extend rendering pipeline. In this first iteration the different passes
are still fixed in the same order as before, but prep and render are
detached and the different passes are responsible for collecting the
data needed in the prep state so it later can be rendered.

Change-Id: I2f9914b335d056dcdde95903f846f18ee0490c75
Reviewed-by: Andy Nichols <andy.nichols@qt.io>