Fix unused variable warning in TinyEXR when building in release mode
The variable is used for an assert() check, which are disabled in
release mode leading to clang warning about an unused variable. This
patch removes the variable and asserts since this is development code
for a 3rd party library.

Change-Id: I251afe1ceeb0030c021fb2e1fed23f791a124e49
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>
(cherry picked from commit 321a2c599325f810e327345141a1e41940aca5ff)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>