Add soft and hard assert checks
These are convenient replacements for regular assert/Q_ASSERT that allow
for wider range of usages. For example, it's often the case that we don't
want an assert to be completely ignored in released code, as an assert
usually means we're hitting an unexpected use-case, so printing a warning
and having a fallback 'action' might be desirable to try to recover. At
the same time it's desirable that those asserts are hard/fatal under
the development cycle to more easily catch mistakes or bugs.

The inspiration for these are pulled from QtC and only put to use in
the passes for now.

Change-Id: Ifbb4506455e785698e91a76299598c9186f35ac5
Reviewed-by: Andy Nichols <andy.nichols@qt.io>