Automatically generate Level of Detail meshes on import
Introduces a new version of the mesh file format that stores
additional levels of detail that are generated by using using
meshoptimizer to simplify the original geometry of each subset.

The simplify process returns a list of new range of index values
that should be used when rendering a mesh at a particular level
of detail, as well as an ideal "distance" which manifests itself
as a ratio of how many pixels will be used to render the mesh.
The idea is that if a model will only be rendered as a small ratio
of screen space, then the full resolution geometry is not needed
to get the same result.

Part of the refinement of the simplification process is to add
additional vertices to lod levels where the attributes such as
vertex normals no longer make sense using the original index. In
this case the attributes are recalculated using the new indexes
of the simplified level of detail.

Currently generation of mesh levels of detail is off by default since
some tweaking may be necessary for the best results. This should instead
be an explicit choice by the user on import via the tooling.

Change-Id: I2a2bc0b75aa30284a40241bdf7b90ff072a0fcac
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>