Backend: Add system for drawing debug lines in world space
This is a debug tool to enable drawing lines and points to annotate
things in world space (scene) coordinates. Currently this is only
usable in the backend and has been used as a visual printf/qdebug
when developing features for the renderer. It may be possible to at
some point plumb this to the front end to enable the user to also
take advantage of this infrastructure, but as we have hacked similar
functionality in before to debug issues, it is a needed development
feature.

Change-Id: Ifbdde5dae556d3f2aab059a381e5b1e01eb8317b
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>