Fix infinite loop when adding same effect twice in SceneEnvironment
Since the effect is using a singly linked list pointing to the next
effect, having the same effect appear twice will lead to a loop in the
list. Now, if the same effect is already added to the list of effects it
will be skipped and a warning is printed.

Pick-to: 6.4
Fixes: QTBUG-102477
Change-Id: Ic9677a488233a39610548671363375beace72bc1
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>