Make aoFactor the definitive Ambient Occlusion value
Previously we applied ao from maps in a differnt way than ao from SSAO.
There really isn't a reason to do that, and SSAO was not having the
desired effect in many cases beause it was applied in so few places.
Now qt_ao is part of qt_aoFactor (where only the qt_ao contribution is
limited by the occlusion factor as defined by ao maps).

This should probably break some lancelot tests.

Change-Id: I02a03fe06fbf6f04bf928fef625153107b883294
Reviewed-by: Christian Strømme <christian.stromme@qt.io>