Hook up DebugView to the QSSGRhiContextStats
Show some of the statistics that are already available.
The important part here is the plumbing, and that the visibility
of the DebugView controls the collection of the data in
QSSGRhiContextStats. (this only works well if there is one DebugView
for one View3D, but that's acceptable I suppose)

Previously the collection of the draw call and render pass data was
a global, static setting based on an env.var. or the profiler being
enabled. That is now complemented with a dynamic flag which the
QQuick3DRenderStats can set based on what it gets from the DebugView.

This now allows us to show the data collected in QSSGRhiContextStats
live, and possibly extend it with more things later on, without
affecting performance when no DebugView is visible.

Also, start naming render target objects properly. Setting the QRhi
object name is a good practice, except if it is on a hot code path
(which should not be common for resource creation).

For render targets in particular the QSSGRhiContextStats will now store
the name from the QRhiRenderTarget.

The full list of render passes is shown only when opted in via a new
renderPassDetailsVisible property.

To make it manageable we use a ScrollView with a limit on the height so
the debug view panel continues to look reasonable.

For lights and effects we want to see the objectName() for the
associated frontend object or in the absence of that the
metaObject->className(). To enable this, there is now a debugObjectName
for any node, effect, or material. The value is generated once upon the
first sync.

Also displays the list of active textures and meshes (registered to the
QSSGRhiContext, so these are not per-View3D but per-window).

In addition, controls for DebugSettings are added as well.

Fixes: QTBUG-106949
Change-Id: Iab6b81b42ed4c2a5bbfd071b67449f1d7fc3409c
Reviewed-by: Andy Nichols <andy.nichols@qt.io>